requirements.txt
Flask==2.2.2
google-cloud-texttospeech==2.12.1